@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
     <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
     <Deterministic>false</Deterministic>
     <IsPackable>false</IsPackable>
@@ -16,7 +16,7 @@
 
   <ItemGroup>
     <Reference Include="Integral">
-      <HintPath>..\Integral\bin\Debug\netstandard2.0\Integral.dll</HintPath>
+      <HintPath>..\Integral\bin\Debug\net5.0\Integral.dll</HintPath>
     </Reference>
   </ItemGroup>
 
